Move TerminalShellIntegration stubs to main theia API
fixes #14107

contributed on behalf of STMicroelectronics
